requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
python-dateutil
python-dotenv
requests
